Checkpoint moderation review changes
