+++ Cargo.toml
@@ -42,6 +42,7 @@ opentelemetry-appender-tracing = "0.31.1"
 askama = "0.15.1"
 async-stream = "0.3"
 base64 = "0.22"
+bitflags = "2"
 bytes = "1"
 cpal = "0.16"
 crossterm = "0.29"
