+++ late-core/src/models/moderation_audit_log.rs
@@ -1,6 +1,6 @@
 use anyhow::Result;
+use deadpool_postgres::GenericClient;
 use serde_json::Value;
-use tokio_postgres::Client;
 use uuid::Uuid;
 
 crate::model! {
@@ -18,23 +18,24 @@ crate::model! {
 
 impl ModerationAuditLog {
     pub async fn record(
-        client: &Client,
+        client: &impl GenericClient,
         actor_user_id: Uuid,
         action: impl Into<String>,
         target_kind: impl Into<String>,
         target_id: Option<Uuid>,
         metadata: Value,
     ) -> Result<Self> {
-        Self::create(
-            client,
-            ModerationAuditLogParams {
-                actor_user_id,
-                action: action.into(),
-                target_kind: target_kind.into(),
-                target_id,
-                metadata,
-            },
-        )
-        .await
+        let action = action.into();
+        let target_kind = target_kind.into();
+        let row = client
+            .query_one(
+                "INSERT INTO moderation_audit_log
+                 (actor_user_id, action, target_kind, target_id, metadata)
+                 VALUES ($1, $2, $3, $4, $5)
+                 RETURNING *",
+                &[&actor_user_id, &action, &target_kind, &target_id, &metadata],
+            )
+            .await?;
+        Ok(Self::from(row))
     }
 }
