+++ late-ssh/src/app/render.rs
@@ -130,7 +130,7 @@ struct DrawContext<'a> {
     activity: &'a std::collections::VecDeque<crate::state::ActivityEvent>,
     banner: Option<&'a Banner>,
     is_admin: bool,
-    is_mod: bool,
+    is_moderator: bool,
     show_right_sidebar: bool,
     show_games_sidebar: bool,
     show_settings: bool,
@@ -411,7 +411,7 @@ impl App {
                         activity: &self.activity,
                         banner: banner.as_ref(),
                         is_admin: self.is_admin,
-                        is_mod: self.is_mod,
+                        is_moderator: self.is_moderator,
                         show_right_sidebar,
                         show_games_sidebar,
                         show_settings: self.show_settings,
@@ -636,7 +636,7 @@ impl App {
                     active_room: ctx.rooms_active_room,
                     blackjack_state: ctx.blackjack_state,
                     is_admin: ctx.is_admin,
-                    is_mod: ctx.is_mod,
+                    is_moderator: ctx.is_moderator,
                     filter: ctx.rooms_filter,
                     search_active: ctx.rooms_search_active,
                     search_query: ctx.rooms_search_query,
