+++ late-ssh/src/app/rooms/ui.rs
@@ -33,7 +33,7 @@ pub struct RoomsPageView<'a> {
     pub active_room: Option<&'a RoomListItem>,
     pub blackjack_state: &'a BlackjackState,
     pub is_admin: bool,
-    pub is_mod: bool,
+    pub is_moderator: bool,
     pub filter: RoomsFilter,
     pub search_active: bool,
     pub search_query: &'a str,
@@ -705,7 +705,7 @@ fn draw_footer(frame: &mut Frame, area: Rect, view: &RoomsPageView<'_>) {
         spans.push(hint_pair("n", "new"));
     }
 
-    if view.is_admin || view.is_mod {
+    if view.is_admin || view.is_moderator {
         spans.push(Span::raw(" · "));
         spans.push(hint_pair("Esc", "back"));
     }
