+++ late-ssh/src/lib.rs
@@ -4,6 +4,7 @@ pub mod authz;
 pub mod config;
 pub mod dartboard;
 pub mod metrics;
+pub mod moderation;
 pub mod session;
 pub mod ssh;
 pub mod state;
