+++ late-ssh/tests/helpers/mod.rs
@@ -267,7 +267,6 @@ pub fn make_app_with_chat_service(
         now_playing_rx: None,
         user_id,
         permissions: Permissions::default(),
-        is_mod: false,
         artboard_banned: false,
         artboard_ban_expires_at: None,
         my_vote: None,
@@ -374,7 +373,6 @@ pub fn make_app_with_paired_client(
         now_playing_rx: None,
         user_id,
         permissions: Permissions::default(),
-        is_mod: false,
         artboard_banned: false,
         artboard_ban_expires_at: None,
         my_vote: None,
